Monitoring enhancement to support on demand jobs and scheduling jobs along with activity log support

+++ src/http/tracker/tracker.service.ts
@@ -1,15 +1,19 @@
-import { Injectable } from '@nestjs/common';
+import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateTrackerDto } from './dto/create-tracker.dto';
 import { Tracker } from 'src/model/tracker.model';
 import { User } from 'src/model/users.model';
+import { MonitorService } from '../monitor/monitor.service';
 
 @Injectable()
 export class TrackerService {
   constructor(
     @InjectRepository(Tracker)
     private trackerRepo: Repository<Tracker>,
+
+    @Inject(forwardRef(() => MonitorService))
+    private monitorService: MonitorService,
   ) {}
 
   findAllByUser(userId: number) {
@@ -24,10 +28,18 @@ export class TrackerService {
       ...createDto,
       user: { id: user.id },
     });
-    return await this.trackerRepo.save(tracker);
+    const output = await this.trackerRepo.save(tracker);
+    await this.monitorService.addOrUpdateCronJob(tracker);
+    return output;
   }
 
   async delete(id: number) {
+    const tracker = await this.trackerRepo.findOne({ where: { id } });
+    if (!tracker) {
+      throw new Error('Tracker not found');
+    }
+    await this.monitorService.removeCronJob(id);
+    await this.trackerRepo.update(id, { is_active: false });
     return await this.trackerRepo.delete(id);
   }
 }
